Fix the wrong counting of records (should exclude deleted ones correctly)

+++ src/main/kotlin/sage/util/PaginationLogic.kt
@@ -1,8 +1,8 @@
 package sage.util
 
 object PaginationLogic {
-  fun pagesCount(size: Int, totalRecords: Long): Int {
-    val division = (totalRecords / size).toInt()
-    return if (totalRecords % size == 0L) division else division + 1
+  fun pagesCount(size: Int, totalRecords: Int): Int {
+    val division = totalRecords / size
+    return if (totalRecords % size == 0) division else division + 1
   }
 }
